If recipe already exists attempt to remove it when bookmark button is pressed
