+++ frontend/pages/recipes.html
@@ -58,7 +58,8 @@
       </div>
     </div>
 
-
+    
+    <script type="module" src="../script/savedrecipes.js"></script>
     <script type="module" src="../script/recipes.js"></script>
     <script type="module" src="../script/recipefinder.js"></script>
     <script type="module" src="../script/user.js"></script>
