+++ frontend/script/recipefinder.js
@@ -1,63 +1,4 @@
-import { user } from "../script/user.js";
-
-const resultsContainer = document.getElementById('results-container')
-
-
-export const toggleSaveRecipe = (recipeID) => {
-    let recipeCard = document.getElementById(recipeID);
-    let bookmarkIcon = recipeCard.querySelector(".bookmark-image"); 
-    if (bookmarkIcon.classList.contains("saved")) {
-        bookmarkIcon.classList.remove("saved");
-        bookmarkIcon.src = "../images/bookmark-off.svg";
-    } else {
-        bookmarkIcon.classList.add("saved");
-        bookmarkIcon.src ="../images/bookmark-on.svg";
-    }
-}
-
-
-export const saveRecipe = async (recipe) => {
-    const token = user.get_token()
-    if (token == null) return
-    
-    const saved_recipe = {
-        id: recipe.id,
-        user_id: null,
-        title: recipe.title,
-        image_url: recipe.image,
-    }
-
-    fetch("http://127.0.0.1:8000/recipes/add", {
-        method: "POST",
-        headers: {
-            "Content-Type": "application/json",
-            "Accept": "application/json",
-            "Authorization": "Bearer " + token
-        },
-        body: JSON.stringify({"recipe": saved_recipe})
-    })
-    .then(response => response.json())
-}
-
-
-export const displayRecipes = (recipes) => {
-    let card; 
-    for (let recipe of recipes) {
-        // Clone base recipe card tempalte
-        card = document.querySelector("div[data-type='template']").cloneNode(true);
-        card.id = recipe.id;
-        card.querySelector("img").src = recipe.image;
-        card.querySelector(".card-title").textContent = recipe.title;
-        card.style.display = "flex";
-        card.querySelector(".save-button").addEventListener('click', async () => {
-            await saveRecipe(recipe)
-            toggleSaveRecipe(recipe.id);
-        });
-
-        resultsContainer.appendChild(card);
-    }
-}
-
+import { displayRecipes } from "../script/recipes.js"
 
 const searchForRecipes = async () => {
     hideFinalView();
